+++ examples/BOOTH1/Blazedemo.java
@@ -11,6 +11,7 @@ import java.io.File;
 import org.openqa.selenium.support.ui.Select;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.*;
 import static org.openqa.selenium.OutputType.*;
 
@@ -20,6 +21,7 @@ public class Blazedemo {
     @Before
     public void setUp() throws Exception {
         wd = new ChromeDriver();
+        //wd = new FirefoxDriver();
         wd.manage().timeouts().implicitlyWait(60, TimeUnit.SECONDS);
     }