+++ site/dat/docs/Changelog.md
@@ -12,6 +12,8 @@
  - support collection-based cloud tests
  - add stubs for `screenshoter` and `capturehar` services
  - add RSpec tests runner for Selenium
+ - add Mocha-based tests runner for Selenium
+ - fix artifact upload in the cloud
 
 ## 1.6.8 <sup>1 sep 2016</sup>
  - fix hamcrest installation for Java-based Selenium tests